namespace move into mypupils
