+++ DfE.GIAP.All/tests/DfE.GIAP.Web.Tests/Controllers/MyPupils/TestDoubles/PupilPresentationOptionsTestDoubles.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using DfE.GIAP.Web.Controllers.MyPupilList.Services.PupilsPresentation.PresenterHandler.Options;
+﻿using DfE.GIAP.Web.Controllers.MyPupilList.Services.PupilsPresentation.PresenterHandler.Options;
 using DfE.GIAP.Web.Controllers.MyPupilList.Services.PupilsPresentation.PresenterHandler.Order;
 using Moq;
 
-namespace DfE.GIAP.Web.Tests.Controllers.MyPupils;
+namespace DfE.GIAP.Web.Tests.Controllers.MyPupils.TestDoubles;
 public static class PupilPresentationOptionsTestDoubles
 {
     public static PupilsPresentationOptions Create(string sortKey) => Create(sortKey, It.IsAny<SortDirection>());
